portalicious: better dto utility type
AB#30486

This addresses the oddity encountered with attribute labels
